Move files into the component folder.
