+++ src/components/NewGame.js
